+++ kernel/build-script/re-build.sh
@@ -35,7 +35,6 @@ done
 
 echo "!!!  Build RPi0 kernel and modules  !!!"
 cd linux-${KERNEL_VERSION}+/
-KERNEL=kernel
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- bcmrpi_defconfig
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- zImage modules dtbs
 cd ..
@@ -44,7 +43,6 @@ echo "-------------------------"
 
 echo "!!!  Build RPi2 kernel and modules  !!!"
 cd linux-${KERNEL_VERSION}-v7+/
-KERNEL=kernel7
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- bcm2709_defconfig
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- zImage modules dtbs
 cd ..
@@ -53,7 +51,6 @@ echo "-------------------------"
 
 echo "!!!  Build RPi3/4 32-bit kernel and modules  !!!"
 cd linux-${KERNEL_VERSION}-v7l+/
-KERNEL=kernel7l
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- bcm2711_defconfig
 make -j${CPU} ARCH=arm CROSS_COMPILE=arm-linux-gnueabihf- zImage modules dtbs
 cd ..
@@ -62,7 +59,6 @@ echo "-------------------------"
 
 echo "!!!  Build RPi3/4/5 64-bit kernel and modules  !!!"
 cd linux-${KERNEL_VERSION}-v8+/
-KERNEL=kernel8
 make -j${CPU} ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- bcm2711_defconfig
 make -j${CPU} ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- Image modules dtbs
 cd ..
@@ -89,7 +85,7 @@ xz -f linux-${KERNEL_VERSION}-v8+/${MODULE_PATH}
 
 echo "!!!  Creating archive  !!!"
 PKG="cst328-rpi-${KERNEL_VERSION}"
-rm -rf ${PKG}/
+rm -rf "${PKG:?}/"
 
 mkdir -p ${PKG}/boot/overlays
 for V in "+" "-v7+" "-v7l+" "-v8+"; do
@@ -103,7 +99,7 @@ cp linux-${KERNEL_VERSION}+/${OVERLAY_DIR}/cst328.dtbo ${PKG}/boot/overlays/
 tar -czf ${PKG}.tar.gz ${PKG}/ --owner=0 --group=0
 md5sum ${PKG}.tar.gz > ${PKG}.md5sum.txt
 sha1sum ${PKG}.tar.gz > ${PKG}.sha1sum.txt
-rm -rf ${PKG}/
+rm -rf "${PKG:?}/"
 mkdir -p ../output
 mv ${PKG}* ../output/
